Clear some internal state of REST Assured between tests
It seems to improve the class loader leaks issues.

(cherry picked from commit d405a304b186f90cbba08b023091f68a8a33dac4)
